refactor(input): split recipe catalog and panel into subcomponents
Extract RecipeCatalogBrowserList/Panels, InputRecipeRail, and
InputModelSourceSection so CodeFactor complexity checks pass on the
orchestrator shells without changing behavior.
